+++ src/global.d.ts
@@ -1,0 +1,3 @@
+declare module 'sql-formatter-plus' {
+  function format(q: string): string;
+}